TreeDetails slider
