+++ index.web.js
@@ -5,6 +5,8 @@
 import {registerRootComponent} from 'expo';
 import App from './App';
 import 'react-toastify/dist/ReactToastify.css';
+import 'slick-carousel/slick/slick.css';
+import 'slick-carousel/slick/slick-theme.css';
 import './index.css';
 
 console.log('running index.web');
