+++ package.json
@@ -38,6 +38,7 @@
     "apollo-link-ethereum-resolver-web3js": "0.1.3",
     "apollo-link-rest": "0.9.0-rc.1",
     "axios": "0.24.0",
+    "expo": "^44.0.6",
     "google-libphonenumber": "3.2.26",
     "graphql": "16.3.0",
     "graphql-anywhere": "4.2.7",
